requirements.txt
Flask==3.0.3
pdfminer.six==20231228
pytesseract==0.3.13
Pillow==10.4.0
vaderSentiment==3.3.2
requests==2.32.3
gunicorn==22.0.0
python-dotenv==1.0.1
python-docx
pdf2image